% File pkg/man/oo.Rd
% $Id: addHandler.Rd 17 2010-04-08 06:04:24Z mariotomo $

\name{oo}
\alias{GnuPG}
\title{create the main encrypting object}
\description{
  invoke GnuPG$new to initialize an object that will provide access to
  all encrypting and decrypting functions.
}
\usage{
obj <- GnuPG$new(homedir)
}

